Move version message to use-init-data
