+++ enterprise/frontend/src/embedding-sdk/hooks/private/use-init-data/use-init-data.ts
@@ -3,6 +3,7 @@ import { useMount } from "react-use";
 import _ from "underscore";
 
 import { getEmbeddingSdkVersion } from "embedding-sdk/config";
+import { isSdkVersionCompatibleWithMetabaseVersion } from "embedding-sdk/lib/version-utils";
 import { useLazySelector } from "embedding-sdk/sdk-package/hooks/private/use-lazy-selector";
 import { initAuth } from "embedding-sdk/store/auth";
 import {
@@ -10,7 +11,10 @@ import {
   setMetabaseClientUrl,
   setMetabaseInstanceVersion,
 } from "embedding-sdk/store/reducer";
-import { getFetchRefreshTokenFn } from "embedding-sdk/store/selectors";
+import {
+  getFetchRefreshTokenFn,
+  getMetabaseInstanceVersion,
+} from "embedding-sdk/store/selectors";
 import type { SdkStore } from "embedding-sdk/store/types";
 import type { MetabaseAuthConfig } from "embedding-sdk/types";
 import { EMBEDDING_SDK_CONFIG } from "metabase/embedding-sdk/config";
@@ -73,10 +77,24 @@ export const useInitData = ({
 
     dispatch(initAuth(authConfig));
 
-    const EMBEDDING_SDK_VERSION = getEmbeddingSdkVersion();
+    const mbVersion = getMetabaseInstanceVersion(reduxStore.getState());
+    const sdkVersion = getEmbeddingSdkVersion();
+
+    if (
+      mbVersion &&
+      sdkVersion !== "unknown" &&
+      !isSdkVersionCompatibleWithMetabaseVersion({ mbVersion, sdkVersion })
+    ) {
+      console.warn(
+        `SDK version ${sdkVersion} is not compatible with MB version ${mbVersion}, this might cause issues.`,
+        // eslint-disable-next-line no-unconditional-metabase-links-render -- This links only shows for admins.
+        "Learn more at https://www.metabase.com/docs/latest/embedding/sdk/version",
+      );
+    }
+
     api.requestClient = {
       name: EMBEDDING_SDK_CONFIG.metabaseClientRequestHeader,
-      version: EMBEDDING_SDK_VERSION,
+      version: sdkVersion,
     };
 
     api.onResponseError = ({
@@ -91,7 +109,7 @@ export const useInitData = ({
       // eslint-disable-next-line no-console
       console.log(
         // eslint-disable-next-line no-literal-metabase-strings -- Not a user facing string
-        `Using Metabase Embedding SDK, version ${EMBEDDING_SDK_VERSION}`,
+        `Using Metabase Embedding SDK, version ${sdkVersion}`,
       );
     }
   });
